Revert "Zero data check"
