+++ web-app/js/portal/cart/WmsInjector.js
@@ -9,17 +9,5 @@ Portal.cart.WmsInjector = Ext.extend(Portal.cart.BaseInjector, {
         });
 
         return describer.buildDescription('<br />') || "";
-    },
-
-    getInjectionJson: function(collection) {
-        var injectionJson = Portal.cart.WmsInjector.superclass.getInjectionJson(collection);
-
-        if ((collection.totalFilteredFeatures != undefined  && collection.totalFilteredFeatures == 0)) {
-            injectionJson.errorMessage = OpenLayers.i18n("subsetRestrictiveFiltersText");
-        }
-
-        injectionJson.dataFilters = this._getDataFilterEntry(collection);
-
-        return injectionJson;
     }
 });
