+++ web-app/js/portal/filter/ui/FilterGroupPanel.js
@@ -5,18 +5,12 @@ Portal.filter.ui.FilterGroupPanel = Ext.extend(Ext.Container, {
 
         this.map = cfg.map;
         this.loadingMessage = this._createLoadingMessageContainer();
-        this.warningEmptyDownloadMessage =  new Portal.common.AlertMessagePanel({
-            message: OpenLayers.i18n('subsetRestrictiveFiltersText'),
-            messageCss: "alert-info"
-        });
-
         var config = Ext.apply({
             autoDestroy: true,
             autoHeight: true,
             cls: 'filterGroupPanel',
             items: [
-                this.loadingMessage,
-                this.warningEmptyDownloadMessage
+                this.loadingMessage
             ]
         }, cfg);
 
@@ -29,7 +23,6 @@ Portal.filter.ui.FilterGroupPanel = Ext.extend(Ext.Container, {
         var filters = this.dataCollection.getFilters();
         if (filters == undefined) {
             this.dataCollection.on(Portal.data.DataCollection.EVENTS.FILTERS_LOAD_SUCCESS, this._filtersLoaded, this);
-            this.dataCollection.on(Portal.data.DataCollection.EVENTS.FILTERS_UPDATED, this.testGetFeaturesWithFilters, this);
             this.dataCollection.on(Portal.data.DataCollection.EVENTS.FILTERS_LOAD_FAILURE, function() { this._filtersLoaded([]); }, this);
         }
         else {
@@ -62,8 +55,15 @@ Portal.filter.ui.FilterGroupPanel = Ext.extend(Ext.Container, {
         });
     },
 
+    _createErrorMessageContainer: function() {
+        return new Ext.Container({
+            autoEl: 'div',
+            html: ""
+        });
+    },
+
     _setErrorMessageText: function(msg, errorMsgContainer) {
-        errorMsgContainer.html = String.format("<div class=\"alert alert-warning\">{0}</div>", msg);
+        errorMsgContainer.html = "<i>" + msg + "</i>";
     },
 
     _removeLoadingMessage: function() {
@@ -121,46 +121,6 @@ Portal.filter.ui.FilterGroupPanel = Ext.extend(Ext.Container, {
         else {
             this._handleFilterLoadFailure();
         }
-    },
-
-    testGetFeaturesWithFilters: function() {
-
-        var wfsFullCheckUrl = this._getFeatureUrlGeneratorFunction();
-        if (wfsFullCheckUrl.includes("CQL_FILTER")) {
-            Ext.Ajax.request({
-                url: Ext.ux.Ajax.constructProxyUrl(wfsFullCheckUrl),
-                scope: this,
-                success: this._handleGetFeatureRequestResults,
-                failure: this._handleGetFeatureRequestResults
-            });
-        }
-        else {
-            delete(this.dataCollection.totalFilteredFeatures);
-            this._handleEmptyDownloadMsg();
-        }
-    },
-
-    _handleGetFeatureRequestResults: function(results) {
-        var res = Ext.util.JSON.decode(results.responseText);
-        this.dataCollection.totalFilteredFeatures = (res && res.totalFeatures >= 0) ? res.totalFeatures: undefined;
-        this._handleEmptyDownloadMsg();
-    },
-
-    _handleEmptyDownloadMsg: function() {
-        var show = (this.dataCollection.totalFilteredFeatures != undefined  && this.dataCollection.totalFilteredFeatures == 0);
-        this.warningEmptyDownloadMessage.setVisible(show);
-    },
-
-    // uses the WMS map layer
-    _getFeatureUrlGeneratorFunction: function() {
-
-        var url =  OpenLayers.Layer.WMS.getFeatureRequestUrl(
-            this.dataCollection.getFilters(),
-            this.dataCollection.layerSelectionModel.selectedLayer.url,
-            this.dataCollection.layerSelectionModel.selectedLayer.wmsName.split('#')[0],
-            "application/json"
-        );
-        return url + "&maxFeatures=1"
     },
 
     _handleFilterLoadFailure: function() {
